contrib/globalsign/mgo: Moved Collection to its own file.
